sockets

+++ services/multimedia_services.go
@@ -4,10 +4,10 @@ import (
 	"errors"
 	"fmt"
 	"github.com/gofiber/fiber/v2"
-	"github.com/nikola43/ecoapigorm/wasabis3manager"
 	database "github.com/nikola43/ecoapigorm/database"
 	"github.com/nikola43/ecoapigorm/models"
 	"github.com/nikola43/ecoapigorm/utils"
+	"github.com/nikola43/ecoapigorm/wasabis3manager"
 	"log"
 	"mime/multipart"
 	"os"
@@ -25,73 +25,109 @@ func UploadMultimedia(
 	uploadMode uint,
 	clinicId uint) error {
 
-	// Save file to root directory:
-	reg, err := regexp.Compile("[^a-zA-Z0-9-.]+")
-	if err != nil {
-		log.Fatal(err)
-	}
+	// sanitize file name
+	reg, _ := regexp.Compile("[^a-zA-Z0-9-.]+")
 	cleanFilename := reg.ReplaceAllString(uploadedFile.Filename, "")
 	clinicName = reg.ReplaceAllString(clinicName, "")
 	clinicName = strings.ToLower(strings.ReplaceAll(clinicName, " ", ""))
-	var clientIDString = strconv.FormatUint(uint64(clientID), 10)
-	os.MkdirAll(fmt.Sprintf("./tempFiles/%s/%s", clinicName, clientIDString), os.ModePerm)
 
-	err = context.SaveFile(uploadedFile, fmt.Sprintf("./tempFiles/%s/%s/%s", clinicName, clientIDString, cleanFilename))
+	clientIDString := strconv.FormatUint(uint64(clientID), 10)
+	fileType := utils.GetFileType(cleanFilename, uploadMode)
+	url := "https://s3.eu-central-1.wasabisys.com/steleros/" + clinicName + "/" + clientIDString + "/" + fileType + "/" + cleanFilename
+
+	//uploadedFilePath := ""
+	clientFolder := "./tempFiles/" + clinicName + "/" + clientIDString
+
+	if _, err := os.Stat(clientFolder); os.IsNotExist(err) {
+		os.MkdirAll(clientFolder, os.ModePerm)
+	}
+
+	err := context.SaveFile(uploadedFile, clientFolder+"/"+cleanFilename)
 	if err != nil {
 		return err
 	}
 
-	fileType := utils.GetFileType(cleanFilename, uploadMode)
+	file, err := os.Open(clientFolder + "/" + cleanFilename)
+	if err != nil {
+		fmt.Printf("err opening file: %s", err)
+	}
+	defer file.Close()
+	fileInfo, _ := file.Stat()
+	size := fileInfo.Size()
+
 	fmt.Println("fileType")
 	fmt.Println(fileType)
 
-	if fileType == "image" {
-		// image
+	if _, err := os.Stat(clientFolder + "/" + fileType); os.IsNotExist(err) {
+		os.MkdirAll(clientFolder+"/"+fileType, os.ModePerm)
+	}
 
-		url, size, storeInAmazonError := wasabis3manager.WasabiS3Client.UploadObject(
-			bucketName,
-			clinicName,
-			"tempFiles/"+clinicName+"/"+clientIDString+"/"+cleanFilename,
-			clientID,
-			fileType,
-		)
-		if storeInAmazonError != nil {
-			fmt.Println(storeInAmazonError.Error())
-		}
+	switch fileType {
+	case "image":
+
+		// image
 		image := models.Image{
-			Filename: cleanFilename,
-			ClientID: clientID,
-			Url:      url,
-			Size:     uint(size),
-			ClinicID: clinicId,
+			Filename:  cleanFilename,
+			ClientID:  clientID,
+			Url:       url,
+			Size:      uint(size),
+			ClinicID:  clinicId,
+			Available: false,
 		}
 		database.GormDB.Create(&image)
 
-		e := os.Remove(fmt.Sprintf("./tempFiles/%s/%s/%s", clinicName, clientIDString, cleanFilename))
-		if e != nil {
-			fmt.Println(e)
-			//log.Fatal(e)
-		}
+		go func() {
+			err = utils.CompressImage("tempFiles/"+clinicName+"/"+clientIDString+"/"+cleanFilename, "tempFiles/"+clinicName+"/"+clientIDString+"/"+fileType+"/"+cleanFilename)
+			if err != nil {
+				fmt.Println(err.Error())
+			}
+
+			imageUrl, imageSize, storeInAmazonError := wasabis3manager.WasabiS3Client.UploadObject(
+				bucketName,
+				clinicName,
+				"tempFiles/"+clinicName+"/"+clientIDString+"/"+fileType+"/"+cleanFilename,
+				clientID,
+				fileType,
+			)
+			if storeInAmazonError != nil {
+				fmt.Println(storeInAmazonError.Error())
+			}
+
+			imageUpdate := new(models.Image)
+			result := database.GormDB.Where("url = ?", imageUrl).Find(&imageUpdate)
+			if result.Error != nil {
+				log.Fatal(result.Error)
+				return
+			}
+
+			database.GormDB.Model(&imageUpdate).Where("url = ?", imageUrl).Update("available", true)
+
+			fmt.Println(imageUrl)
+			fmt.Println(imageSize)
+
+			/*
+				e := os.Remove("tempFiles/" + clinicName + "/" + clientIDString + "/" + cleanFilename)
+				if e != nil {
+					//fmt.Println(e)
+					panic(e)
+				}
+				e = os.Remove(fmt.Sprintf("tempFiles/" + clinicName + "/" + clientIDString + "/" + fileType + "/" + cleanFilename)
+				if e != nil {
+					//fmt.Println(e)
+					panic(e)
+				}
+			*/
+		}()
 
 		return err
-	} else if fileType == "video" || fileType == "holographic" {
-		// upload video
-		videoUrl, videoSize, storeInAmazonError := wasabis3manager.WasabiS3Client.UploadObject(
-			bucketName,
-			clinicName,
-			"tempFiles/"+clinicName+"/"+clientIDString+"/"+cleanFilename,
-			clientID,
-			fileType,
-		)
-		if storeInAmazonError != nil {
-			return err
-		}
+		break
+	case "video":
 
 		ThumbnailCleanFilename := "tempFiles/" + clinicName + "/" + clientIDString + "/" + cleanFilename + "_thumbnail.jpg"
 		ThumbnailCleanFilename = strings.ToLower(strings.ReplaceAll(ThumbnailCleanFilename, " ", ""))
 
 		// create thumb
-		thumbnailPath, extractThumbnailFromVideoError := utils.ExtractThumbnailFromVideo("tempFiles/" + clinicName + "/" + clientIDString + "/" + cleanFilename)
+		extractThumbnailFromVideoError := utils.ExtractThumbnailFromVideo("tempFiles/"+clinicName+"/"+clientIDString+"/"+cleanFilename, "tempFiles/"+clinicName+"/"+clientIDString+"/"+fileType+"/"+cleanFilename+"_thumbnail.jpg")
 		if extractThumbnailFromVideoError != nil {
 			return extractThumbnailFromVideoError
 		}
@@ -100,7 +136,7 @@ func UploadMultimedia(
 		thumbUrl, thumbSize, storeThumbInAmazonError := wasabis3manager.WasabiS3Client.UploadObject(
 			bucketName,
 			clinicName,
-			thumbnailPath,
+			"tempFiles/"+clinicName+"/"+clientIDString+"/"+fileType+"/"+cleanFilename+"_thumbnail.jpg",
 			clientID,
 			fileType,
 		)
@@ -112,9 +148,9 @@ func UploadMultimedia(
 			video := models.Video{
 				Filename:     cleanFilename,
 				ClientID:     clientID,
-				Url:          videoUrl,
+				Url:          url,
 				ThumbnailUrl: thumbUrl,
-				Size:         uint(videoSize + thumbSize),
+				Size:         uint(size + thumbSize),
 				ClinicID:     clinicId,
 			}
 			database.GormDB.Create(&video)
@@ -124,37 +160,69 @@ func UploadMultimedia(
 			video := models.Holographic{
 				Filename:     cleanFilename,
 				ClientID:     clientID,
-				Url:          videoUrl,
+				Url:          url,
 				ThumbnailUrl: thumbUrl,
-				Size:         uint(videoSize + thumbSize),
+				Size:         uint(size + thumbSize),
 				ClinicID:     clinicId,
 			}
 			database.GormDB.Create(&video)
 		}
 
-		e := os.Remove(fmt.Sprintf("./tempFiles/%s/%s/%s", clinicName, clientIDString, cleanFilename))
-		if e != nil {
-			fmt.Println(e)
-		}
-
-		e = os.Remove(thumbnailPath)
-		if e != nil {
-			fmt.Println(e)
-		}
-
-		return err
-	} else if fileType == "holo" {
-		// holo
-		/*
-			url, size, storeInAmazonError := utils.UploadObject("tempFiles/"+cleanFilename, clientID, fileType)
-			if storeInAmazonError != nil {
-				fmt.Println(storeInAmazonError.Error())
+		go func() {
+			fmt.Println("fileComrpess")
+			err := utils.CompressMP4("tempFiles/" + clinicName + "/" + clientIDString + "/" + cleanFilename, "tempFiles/"+clinicName+"/"+clientIDString+"/"+fileType+"/"+cleanFilename)
+			if err != nil {
+				log.Fatal(err)
+				return
 			}
-			image := models.Video{ClientID: clientID, Url: url, Size: uint(size)}
-			database.GormDB.Create(&image)
-		*/
+
+			// upload video
+			videoUrl, videoSize, storeInAmazonError := wasabis3manager.WasabiS3Client.UploadObject(
+				bucketName,
+				clinicName,
+				"tempFiles/"+clinicName+"/"+clientIDString+"/"+fileType+"/"+cleanFilename,
+				clientID,
+				fileType,
+			)
+			if storeInAmazonError != nil {
+				log.Fatal(storeInAmazonError)
+				return
+			}
+
+			fmt.Println(videoUrl)
+			fmt.Println(videoSize)
+
+			videoUpdate := new(models.Image)
+			result := database.GormDB.Where("url = ?", videoUrl).Find(&videoUpdate)
+			if result.Error != nil {
+				log.Fatal(result.Error)
+				return
+			}
+
+			database.GormDB.Model(&videoUpdate).Where("url = ?", videoUrl).Update("available", true)
+
+			fmt.Println("fin")
+
+
+
+			/*
+			e := os.Remove(fmt.Sprintf("./tempFiles/%s/%s/%s", clinicName, clientIDString, cleanFilename))
+			if e != nil {
+				fmt.Println(e)
+			}
+
+			e = os.Remove("tempFiles/"+clinicName+"/"+clientIDString+"/"+fileType+"/"+cleanFilename+"_thumbnail.jpg")
+			if e != nil {
+				fmt.Println(e)
+			}
+			*/
+
+		}()
+
 		return err
-	} else if fileType == "heartbeat" {
+		break
+	case "heartbeat":
+
 		// holo
 		url, size, storeInAmazonError := wasabis3manager.WasabiS3Client.UploadObject(bucketName, clinicName, "tempFiles/"+clinicName+"/"+clientIDString+"/"+cleanFilename, clientID, fileType)
 		if storeInAmazonError != nil {
@@ -169,9 +237,10 @@ func UploadMultimedia(
 		}
 
 		return err
-	} else {
-		return errors.New("invalid file")
+		break
 	}
+
+	return errors.New("invalid file")
 }
 
 func DeleteImage(bucketName string, imageID uint) error {
